requirements.txt
opencv-python
pillow
pandas
numpy
tensorflow
scikit-learn
streamlit